handle arbitrary map dimensions + fix bug with asymmetric PSFs
I was stupidly thinking that the transpose of the PSF is the
same as reflecting the PSF over both axes. Instead of doing these
flips in real space I am now just taking the complex conjugate
of the FFT of the PSF (this is so that we can use a fast convolution
to obtain the cross-correlation).